Lotman: select libLotMan C ABI to match the deployed XRootD purge plugin
The XRootD pfc purge plugin (libXrdPurgeLotMan) links libLotMan at a fixed
C ABI. lotman v0.1.0 added query_time / include_reclaimed / hierarchical
parameters to the eviction queries and update_lot_usage_by_dir; a plugin
built against the older v0.0.4 ABI passes fewer arguments, so a new-ABI
libLotMan reads shifted registers and segfaults while converting the
returned lot list during purge.

Build libLotMan at whichever ABI matches the deployed plugin:

- Split the version-sensitive exports into export_api_v1.go (current ABI,
  the default) and export_api_legacy.go (old v0.0.4 signatures), selected by
  the lotman_legacy_api build tag. Shared JSON parsing moves to parseDirUsage
  in export.go.
- Makefile lotman-shared resolves LOTMAN_ABI (auto|new|legacy, default auto):
  auto sniffs the installed lotman header (new iff update_lot_usage_by_dir
  takes an int64_t query_time) and adds -tags lotman_legacy_api when the old
  ABI is in use; override with LOTMAN_ABI=new|legacy.

Verified: a legacy-ABI libLotMan plus the stock old purge plugin runs the V1
cache eviction path with no crash and real lot-based eviction.
